move InterruptType into cpu module
