+++ ch22-core/src/address_map/io_space/io_device_list.rs
@@ -1,8 +1,8 @@
 use std::collections::HashMap;
 
 use crate::address_map::io_space::DeviceSpeed;
+use crate::cpu::InterruptType;
 use crate::devices::IODevice;
-use crate::interrupt_type::InterruptType;
 use crate::word::Word;
 
 pub type IODeviceID = usize;
