+++ ch22-core/src/cpu/cpu_io/cpu_io_mock.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::cpu::cpu_io::CpuIO;
-use crate::interrupt_type::InterruptType;
+use crate::cpu::interrupt_type::InterruptType;
 use crate::word::Word;
 
 #[derive(Default)]
