+++ ch22-core/src/cpu/executor.rs
@@ -3,12 +3,11 @@ mod addressing;
 mod memory_util;
 mod unary_ops;
 
-use crate::interrupt_type::InterruptType;
-use crate::word::Word;
-
 use super::cpu_io::CpuIO;
 use super::interrupt_due_state::InterruptDueState;
+use super::interrupt_type::InterruptType;
 use super::registers::*;
+use crate::word::Word;
 
 use accumulator_binary_ops::*;
 use addressing::*;
