+++ ch22-core/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod address_map;
 pub mod cpu;
 mod devices;
-pub mod interrupt_type;
 pub mod system;
 mod utils;
 pub mod video;
